requirements.txt
asgiref==3.5.2
async-generator==1.10
attrs==22.1.0
backports.zoneinfo==0.2.1
certifi==2022.9.24
cffi==1.15.1
charset-normalizer==2.1.1
cryptography==38.0.1
defusedxml==0.7.1
Django==4.1.1
django-allauth==0.51.0
django-crispy-forms==1.14.0
django-embed-video==1.4.5
exceptiongroup==1.0.0rc9
h11==0.14.0
idna==3.4
oauthlib==3.2.1
outcome==1.2.0
pycparser==2.21
PyJWT==2.5.0
PySocks==1.7.1
python-dotenv==0.21.0
python3-openid==3.2.0
requests==2.28.1
requests-oauthlib==1.3.1
selenium==4.5.0
sniffio==1.3.0
sortedcontainers==2.4.0
sqlparse==0.4.3
tqdm==4.64.1
trio==0.22.0
trio-websocket==0.9.2
types-cryptography==3.3.23
urllib3==1.26.12
webdriver-manager==3.8.3
wsproto==1.2.0
